+++ src/Tuc/ParsedTypes.fs
@@ -153,10 +153,13 @@ and ParsedKeyWordWithBody<'Type> = {
 
 and ParsedKeyWordIf<'Type> = {
     Value: 'Type
+    IfKeyWord: KeyWord
     IfLocation: ParsedLocation
     ConditionLocation: ParsedLocation
-    ElseLocation: ParsedLocation option
     Body: Parsed<TucPart> list
+
+    ElseKeyWord: KeyWord option
+    ElseLocation: ParsedLocation option
     ElseBody: (Parsed<TucPart> list) option
 }
 
@@ -239,7 +242,7 @@ module Parsed =
         | Parsed.KeyWord k -> Parsed.KeyWord { KeyWord = k.KeyWord; KeyWordLocation = k.KeyWordLocation; ValueLocation = k.ValueLocation; Value = k.Value |> f }
         | Parsed.KeyWordWithoutValue k -> Parsed.KeyWordWithoutValue { KeyWord = k.KeyWord; KeyWordLocation = k.KeyWordLocation }
         | Parsed.KeyWordWithBody k -> Parsed.KeyWordWithBody { KeyWord = k.KeyWord; KeyWordLocation = k.KeyWordLocation; ValueLocation = k.ValueLocation; Body = k.Body; Value = k.Value |> f }
-        | Parsed.KeyWordIf k -> Parsed.KeyWordIf { IfLocation = k.IfLocation; ConditionLocation = k.ConditionLocation; ElseLocation = k.ElseLocation; Body = k.Body; ElseBody = k.ElseBody; Value = k.Value |> f }
+        | Parsed.KeyWordIf k -> Parsed.KeyWordIf { IfKeyWord = k.IfKeyWord; IfLocation = k.IfLocation; ConditionLocation = k.ConditionLocation; ElseKeyWord = k.ElseKeyWord; ElseLocation = k.ElseLocation; Body = k.Body; ElseBody = k.ElseBody; Value = k.Value |> f }
         | Parsed.ParticipantDefinition p -> Parsed.ParticipantDefinition { Context = p.Context; Domain = p.Domain; Alias = p.Alias; Value = p.Value |> f }
         | Parsed.ComponentDefinition c -> Parsed.ComponentDefinition { Context = c.Context; Domain = c.Domain; Participants = c.Participants; Value = c.Value |> f }
         | Parsed.Lifeline l -> Parsed.Lifeline { ParticipantLocation = l.ParticipantLocation; Execution = l.Execution; Value = l.Value |> f }